requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
gradio
requests
apscheduler
google-cloud-storage
google-cloud-bigquery
lightgbm
